+++ internal/provisioner/helm/nginx_internal.go
@@ -2,7 +2,7 @@
 // See LICENSE.txt for license information.
 //
 
-package provisioner
+package helm
 
 import (
 	"fmt"
@@ -15,10 +15,12 @@ import (
 )
 
 const (
-	namespaceNginx = "nginx"
+	namespaceNginxInternal           = "nginx-internal"
+	chartDeploymentNameNginxInternal = "nginx-internal"
+	chartNameNginxInternal           = "ingress-nginx/ingress-nginx"
 )
 
-type nginx struct {
+type nginxInternal struct {
 	awsClient      aws.AWS
 	kubeconfigPath string
 	logger         log.FieldLogger
@@ -28,33 +30,33 @@ type nginx struct {
 	provisioner    string
 }
 
-func newNginxHandle(version *model.HelmUtilityVersion, cluster *model.Cluster, kubeconfigPath string, awsClient aws.AWS, logger log.FieldLogger) (*nginx, error) {
+func NewNginxInternalHandle(version *model.HelmUtilityVersion, cluster *model.Cluster, kubeconfigPath string, awsClient aws.AWS, logger log.FieldLogger) (*nginxInternal, error) {
 	if logger == nil {
-		return nil, errors.New("cannot instantiate NGINX handle with nil logger")
+		return nil, errors.New("cannot instantiate NGINX INTERNAL handle with nil logger")
 	}
+
 	if cluster == nil {
-		return nil, errors.New("cannot create a connection to Nginx if the cluster provided is nil")
+		return nil, errors.New("cannot create a connection to Nginx internal if the cluster provided is nil")
 	}
 	if awsClient == nil {
-		return nil, errors.New("cannot create a connection to Nginx if the awsClient provided is nil")
+		return nil, errors.New("cannot create a connection to Nginx internal if the awsClient provided is nil")
 	}
 	if kubeconfigPath == "" {
 		return nil, errors.New("cannot create utility without kubeconfig")
 	}
 
-	return &nginx{
+	return &nginxInternal{
 		awsClient:      awsClient,
 		kubeconfigPath: kubeconfigPath,
 		cluster:        cluster,
-		logger:         logger.WithField("cluster-utility", model.NginxCanonicalName),
+		logger:         logger.WithField("cluster-utility", model.NginxInternalCanonicalName),
 		desiredVersion: version,
-		actualVersion:  cluster.UtilityMetadata.ActualVersions.Nginx,
+		actualVersion:  cluster.UtilityMetadata.ActualVersions.NginxInternal,
 		provisioner:    cluster.Provisioner,
 	}, nil
-
 }
 
-func (n *nginx) updateVersion(h *helmDeployment) error {
+func (n *nginxInternal) updateVersion(h *helmDeployment) error {
 	actualVersion, err := h.Version()
 	if err != nil {
 		return err
@@ -64,10 +66,10 @@ func (n *nginx) updateVersion(h *helmDeployment) error {
 	return nil
 }
 
-func (n *nginx) CreateOrUpgrade() error {
-	h, err := n.NewHelmDeployment()
+func (n *nginxInternal) CreateOrUpgrade() error {
+	h, err := n.NewHelmDeployment(true)
 	if err != nil {
-		return errors.Wrap(err, "failed to generate nginx helm deployment")
+		return errors.Wrap(err, "failed to generate nginx internal helm deployment")
 	}
 
 	err = h.Update()
@@ -77,6 +79,7 @@ func (n *nginx) CreateOrUpgrade() error {
 
 	if err = n.updateVersion(h); err != nil {
 		return err
+
 	}
 
 	if err = n.addLoadBalancerNameTag(); err != nil {
@@ -86,25 +89,25 @@ func (n *nginx) CreateOrUpgrade() error {
 	return nil
 }
 
-func (n *nginx) addLoadBalancerNameTag() error {
+func (n *nginxInternal) addLoadBalancerNameTag() error {
 
-	endpoint, elbType, err := getElasticLoadBalancerInfo(namespaceNginx, n.logger, n.kubeconfigPath)
+	endpoint, elbType, err := getElasticLoadBalancerInfo(namespaceNginxInternal, n.logger, n.kubeconfigPath)
 	if err != nil {
-		return errors.Wrap(err, "couldn't get the loadbalancer endpoint (nginx)")
+		return errors.Wrap(err, "couldn't get the loadbalancer endpoint (nginx-internal)")
 	}
 
 	if err := addLoadBalancerNameTag(n.awsClient.GetLoadBalancerAPIByType(elbType), endpoint); err != nil {
-		return errors.Wrap(err, "failed to add loadbalancer name tag (nginx)")
+		return errors.Wrap(err, "failed to add loadbalancer name tag (nginx-internal)")
 	}
 
 	return nil
 }
 
-func (n *nginx) DesiredVersion() *model.HelmUtilityVersion {
+func (n *nginxInternal) DesiredVersion() *model.HelmUtilityVersion {
 	return n.desiredVersion
 }
 
-func (n *nginx) ActualVersion() *model.HelmUtilityVersion {
+func (n *nginxInternal) ActualVersion() *model.HelmUtilityVersion {
 	if n.actualVersion == nil {
 		return nil
 	}
@@ -115,37 +118,38 @@ func (n *nginx) ActualVersion() *model.HelmUtilityVersion {
 	}
 }
 
-func (n *nginx) Destroy() error {
-	helm, err := n.NewHelmDeployment()
+func (n *nginxInternal) Destroy() error {
+	helm, err := n.NewHelmDeployment(false)
 	if err != nil {
 		return err
 	}
 	return helm.Delete()
 }
 
-func (n *nginx) ValuesPath() string {
+func (n *nginxInternal) ValuesPath() string {
 	if n.desiredVersion == nil {
 		return ""
 	}
 	return n.desiredVersion.Values()
 }
 
-func (n *nginx) Migrate() error {
+func (n *nginxInternal) Migrate() error {
 	return nil
 }
 
-func (n *nginx) NewHelmDeployment() (*helmDeployment, error) {
-	certificate, err := n.awsClient.GetCertificateSummaryByTag(aws.DefaultInstallCertificatesTagKey, aws.DefaultInstallCertificatesTagValue, n.logger)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to retrive the AWS ACM")
-	}
+func (n *nginxInternal) NewHelmDeployment(withArguments bool) (*helmDeployment, error) {
+	var setArguments []string
+	if withArguments {
+		certificate, err := n.awsClient.GetCertificateSummaryByTag(aws.DefaultInstallPrivateCertificatesTagKey, aws.DefaultInstallPrivateCertificatesTagValue, n.logger)
+		if err != nil {
+			return nil, errors.Wrap(err, "failed to retrive the AWS Private ACM")
+		}
 
-	if certificate.ARN == nil {
-		return nil, errors.New("retrieved certificate does not have ARN")
-	}
+		if certificate.ARN == nil {
+			return nil, errors.New("retrieved certificate does not have ARN")
+		}
 
-	setArguments := []string{
-		fmt.Sprintf("controller.service.annotations.service\\.beta\\.kubernetes\\.io/aws-load-balancer-ssl-cert=%s", *certificate.ARN),
+		setArguments = append(setArguments, fmt.Sprintf("controller.service.annotations.service\\.beta\\.kubernetes\\.io/aws-load-balancer-ssl-cert=%s", *certificate.ARN))
 	}
 
 	if n.provisioner == model.ProvisionerEKS {
@@ -154,7 +158,6 @@ func (n *nginx) NewHelmDeployment() (*helmDeployment, error) {
 		// As a workaround, trusted pods that require control plane nodes to connect to them,
 		// such as those implementing admission controller webhooks, can include hostNetwork:true in their pod spec.
 		// See https://docs.tigera.io/calico/3.25/getting-started/kubernetes/managed-public-cloud/eks
-
 		// setArguments = append(setArguments, "controller.hostNetwork=true")
 
 		// hostNetwork can cause port conflict, that's why we need to use DaemonSet
@@ -164,9 +167,9 @@ func (n *nginx) NewHelmDeployment() (*helmDeployment, error) {
 	}
 
 	return newHelmDeployment(
-		"ingress-nginx/ingress-nginx",
-		"nginx",
-		namespaceNginx,
+		chartNameNginxInternal,
+		chartDeploymentNameNginxInternal,
+		namespaceNginxInternal,
 		n.kubeconfigPath,
 		n.desiredVersion,
 		strings.Join(setArguments, ","),
@@ -174,6 +177,6 @@ func (n *nginx) NewHelmDeployment() (*helmDeployment, error) {
 	), nil
 }
 
-func (n *nginx) Name() string {
-	return model.NginxCanonicalName
+func (n *nginxInternal) Name() string {
+	return model.NginxInternalCanonicalName
 }
