+++ internal/provisioner/helm/teleport.go
@@ -2,7 +2,7 @@
 // See LICENSE.txt for license information.
 //
 
-package provisioner
+package helm
 
 import (
 	"fmt"
@@ -24,7 +24,7 @@ type teleport struct {
 	actualVersion  *model.HelmUtilityVersion
 }
 
-func newTeleportHandle(cluster *model.Cluster, desiredVersion *model.HelmUtilityVersion, kubeconfigPath string, awsClient aws.AWS, logger log.FieldLogger) (*teleport, error) {
+func NewTeleportHandle(cluster *model.Cluster, desiredVersion *model.HelmUtilityVersion, kubeconfigPath string, awsClient aws.AWS, logger log.FieldLogger) (*teleport, error) {
 	if logger == nil {
 		return nil, errors.New("cannot instantiate Teleport handle with nil logger")
 	}
