+++ internal/provisioner/kops_utils.go
@@ -7,7 +7,6 @@ package provisioner
 import (
 	"context"
 	"fmt"
-	"strings"
 	"time"
 
 	"github.com/mattermost/mattermost-cloud/internal/tools/kops"
@@ -85,80 +84,6 @@ func waitForNamespacesDeleted(ctx context.Context, namespaces []string, k8sClien
 			time.Sleep(5 * time.Second)
 		}
 	}
-}
-
-// getPrivateLoadBalancerEndpoint returns the private load balancer endpoint of the NGINX service.
-func getPrivateLoadBalancerEndpoint(ctx context.Context, namespace string, logger log.FieldLogger, configPath string) (string, error) {
-	k8sClient, err := k8s.NewFromFile(configPath, logger)
-	if err != nil {
-		return "", err
-	}
-
-	for {
-		services, err := k8sClient.Clientset.CoreV1().Services(namespace).List(context.TODO(), metav1.ListOptions{})
-		if err != nil {
-			return "", err
-		}
-		for _, service := range services.Items {
-			if service.Spec.Type == "LoadBalancer" || strings.HasSuffix(service.Name, "query") {
-				if service.Status.LoadBalancer.Ingress != nil {
-					endpoint := service.Status.LoadBalancer.Ingress[0].Hostname
-					if endpoint == "" {
-						return "", errors.New("loadbalancer endpoint value is empty")
-					}
-
-					return endpoint, nil
-				}
-			}
-		}
-		select {
-		case <-ctx.Done():
-			return "", errors.Wrap(ctx.Err(), "timed out waiting for internal load balancer to become ready")
-		case <-time.After(5 * time.Second):
-		}
-	}
-}
-
-// getElasticLoadBalancerInfo returns the private load balancer endpoint and type of the NGINX service.
-func getElasticLoadBalancerInfo(namespace string, logger log.FieldLogger, configPath string) (string, string, error) {
-	k8sClient, err := k8s.NewFromFile(configPath, logger)
-	if err != nil {
-		return "", "", err
-	}
-
-	services, err := k8sClient.Clientset.CoreV1().Services(namespace).List(context.TODO(), metav1.ListOptions{})
-	if err != nil {
-		return "", "", err
-	}
-
-	for _, service := range services.Items {
-		if service.Spec.Type == "LoadBalancer" {
-			if service.Status.LoadBalancer.Ingress != nil {
-				endpoint := service.Status.LoadBalancer.Ingress[0].Hostname
-				if endpoint != "" {
-					return endpoint, service.Annotations["service.beta.kubernetes.io/aws-load-balancer-type"], nil
-				}
-			}
-		}
-	}
-
-	return "", "", nil
-}
-
-// GetPublicLoadBalancerEndpoint returns the public load balancer endpoint of the NGINX service.
-func (provisioner Provisioner) GetPublicLoadBalancerEndpoint(cluster *model.Cluster, namespace string) (string, error) {
-
-	logger := provisioner.logger.WithFields(log.Fields{
-		"cluster":         cluster.ID,
-		"nginx-namespace": namespace,
-	})
-
-	configLocation, err := provisioner.getClusterKubecfg(cluster)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get kube config path")
-	}
-
-	return getPublicLoadBalancerEndpoint(configLocation, namespace, logger)
 }
 
 func updateKopsInstanceGroupAMIs(kops *kops.Cmd, kopsMetadata *model.KopsMetadata, logger log.FieldLogger) error {
