+++ internal/provisioner/pgbouncer/pgbouncer_test.go
@@ -1,4 +1,4 @@
-package provisioner
+package pgbouncer
 
 import (
 	"fmt"
@@ -51,7 +51,7 @@ func TestGeneratePGBouncerBaseIni(t *testing.T) {
 	config := NewPGBouncerConfig(5, 10, 54, 2000, 63, 11, 44, 0)
 	require.NoError(t, config.Validate())
 
-	ini := config.generatePGBouncerBaseIni()
+	ini := config.GeneratePGBouncerBaseIni()
 	assert.Contains(t, ini, "[pgbouncer]")
 
 	// Most of the other values are integers so just spot check a few
